(ref): rename ForecastCity for city_forecast
